requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
gunicorn==20.1.0
google-cloud-bigquery>=3.20.0
google-cloud-aiplatform>=1.38.0
